Conflicts:
	RhythmSelectionDialog/nbproject/genfiles.properties
	SS_Editor/nbproject/genfiles.properties
	Upgrade/src/org/jjazz/upgrade/UpgradeManager.java

+++ ItemRenderer/src/org/jjazz/ui/itemrenderer/IR_ChordSymbol.java
@@ -200,7 +200,7 @@ public class IR_ChordSymbol extends ItemRenderer implements IR_Copiable
         // Create the TextLayout to get its dimension       
         TextLayout textLayout = new TextLayout(attChordString.getIterator(), frc);
         chordSymbolWidth = (int) TextLayoutUtils.getWidth(textLayout, strChord2, false);
-        chordSymbolHeight = (int) TextLayoutUtils.getHeight(textLayout, frc);
+        chordSymbolHeight = TextLayoutUtils.getHeight(textLayout, frc);
         Insets in = getInsets();
         final int PADDING = 1;
         int wFinal = chordSymbolWidth + 2 * PADDING + in.left + in.right; //  + (needOptionDots(ecs) ? CORNER_SIZE : 0);
